Loader: restructure as dir
