+++ src/components/Loader/__style.js
@@ -1,0 +1,8 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+
+    /* styles */
+
+});
+
