+++ src/screens/SearchResultScreen.js
@@ -26,7 +26,7 @@ import {
 import headerConfig from '../helpers/headerConfig';
 import constants from '../helpers/constants';
 import PersonInfo from '../components/Person/PersonInfo';
-import Loader from '../components/Loader/Loader';
+import Loader from '../components/Loader/index.jsx';
 import authHelpers from '../helpers/authHelpers';
 import RegisterModalsContainer from './../components/auth/RegisterModalsContainer/index.tsx';
 import { ConfirmationModal } from '../components/Person/ConfirmationModal';
